+++ NavMap/app/src/main/res/values/strings.xml
@@ -52,6 +52,7 @@
     <string name="address">Address</string>
     <string name="no_bookmarks">No bookmarks...</string>
     <string name="bookmark_something_and_find_it_here">Bookmark something and find it here!</string>
+    <string name="enter_the_bookmark_name">Enter the bookmark name</string>
 
     <string-array name="landmarksTypes">
         <item>Any</item>
